requirements.txt
SQLAlchemy~=1.4.36
passlib~=1.7.4
psycopg2~=2.9.3
Flask~=2.1.2
kafka-python
sqlalchemy-utils
flask_jwt_extended